refactor: Move App store components to dedicated page-component
